src/dashboard/apigateway/requirements_dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --no-annotate -o apigateway/requirements_dev.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
apigw-manager==4.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backports-tarfile==1.2.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
beautifulsoup4==4.13.4
billiard==4.2.1
bk-crypto-python-sdk==2.0.1
bk-notice-sdk==1.4.2
bkapi-bk-apigateway==1.0.11
bkapi-client-core==1.2.0
bkapi-client-generator==0.1.30
bkapi-component-open==1.1.0
bkpaas-auth==3.1.2
blue-krill==2.1.2
cachetools==6.1.0
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
concurrent-log-handler==0.9.28
coverage==7.9.1
cron-descriptor==1.4.5
cryptography==45.0.4
curlify==2.2.1
dacite==1.9.2
decorator==5.2.1
deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
django==4.2.23
django-add-default-value==0.10.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-dynamic-fixture==4.0.1
django-environ==0.12.0
django-filter==25.1
django-nose==1.4.7
django-prometheus==2.4.0
django-timezone-field==7.1
djangoql==0.18.1
djangorestframework==3.16.0
docutils==0.21.2
drf-yasg==1.21.10
elasticsearch==7.7.1
elasticsearch-dsl==7.4.1
etcd3==0.12.0
exceptiongroup==1.3.0
execnet==2.1.1
executing==2.2.0
faker==30.8.1
fakeredis==2.30.1
filelock==3.18.0
future==1.0.0
gevent==25.5.1
googleapis-common-protos==1.59.1
greenlet==3.2.3
grimp==3.9
grpcio==1.44.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
identify==2.6.12
idna==3.10
import-linter==2.3
importlib-metadata==8.4.0
inflection==0.5.1
iniconfig==2.1.0
ipython==8.29.0
jaraco-classes==3.4.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
jaraco-context==6.0.1 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
jaraco-functools==4.2.1 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
jedi==0.19.2
jeepney==0.9.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'linux'
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonfield==3.1.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
keyring==25.6.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
kombu==5.5.4
lazy-object-proxy==1.11.0
mailbits==0.2.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.7.0 ; platform_machine != 'ppc64le' and platform_machine != 's390x'
mypy==1.16.1
mypy-extensions==1.1.0
nh3==0.2.21
nodeenv==1.9.1
nose==1.3.7
openai==1.88.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.27.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-celery==0.48b0
opentelemetry-instrumentation-dbapi==0.48b0
opentelemetry-instrumentation-django==0.48b0
opentelemetry-instrumentation-logging==0.48b0
opentelemetry-instrumentation-redis==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-wsgi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==25.0
parso==0.8.4
pathable==0.4.4
pathspec==0.12.1
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pillow==11.0.0
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
portalocker==3.2.0
prance==23.6.21.0
pre-commit==4.2.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
protobuf==3.20.3
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pure-eval==0.2.3
py-cpuinfo==9.0.0
py-openapi-schema-to-json-schema==0.0.3
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.18.0
pyjwt==2.9.0
pymysql==1.1.1
pypi-simple==1.6.1
pyproject-api==1.9.1
pytest==8.4.1
pytest-benchmark==5.1.0
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-mock==3.14.1
pytest-pretty==1.3.0
pytest-xdist==3.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-editor==1.0.4
python-json-logger==3.3.0
python-redis-lock==4.0.0
pytoolconfig==1.3.1
pytz==2025.2
pywin32==310 ; sys_platform == 'win32'
pywin32-ctypes==0.2.3 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'win32'
pyyaml==6.0.2
raven==6.10.0
readme-renderer==44.0
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
responses==0.25.7
rfc3339-validator==0.1.4
rfc3986==2.0.0
rich==14.0.0
rope==1.13.0
rpds-py==0.25.1
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.12 ; platform_python_implementation == 'CPython'
ruff==0.12.0
secretstorage==3.3.3 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'linux'
setuptools==75.3.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
thrift==0.22.0
toml==0.10.2
tomli==2.2.1
tongsuopy-crayon==1.0.2b6
tox==4.23.2
tqdm==4.67.1
traitlets==5.14.3
twine==6.1.0
types-cachetools==6.0.0.20250525
types-cffi==1.17.0.20250523
types-docutils==0.21.0.20250604
types-pymysql==1.1.0.20240524
types-pyopenssl==24.1.0.20240722
types-python-dateutil==2.9.0.20250516
types-pytz==2025.2.0.20250516
types-pyyaml==6.0.12.20240311
types-redis==4.6.0.20241004
types-requests==2.31.0.1
types-setuptools==80.9.0.20250529
types-urllib3==1.26.25.14
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
virtualenv==20.31.2
watchdog==6.0.0
wcwidth==0.2.13
werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
zipp==3.23.0
zope-event==5.0
zope-interface==7.2
